chore(rsc2): expand requests
